Convert MapHelper from singleton class to static helper

+++ src/main/java/datagenerator/CarbonDioxideDataGenerator.java
@@ -18,7 +18,7 @@ public class CarbonDioxideDataGenerator implements SensorDataGenerator {
         random = new Random(DEFAULT_SEED);
     }
 
-    public static Double generateData(double x, double y) {
+    public static double generateData(double x, double y) {
         if (x < 200 && y < 230)
             return (double) 97 - 20 + (x + y) / 250 + 0.3 * random.nextGaussian();
         else if (x < 1000 && y < 1000)
@@ -38,8 +38,8 @@ public class CarbonDioxideDataGenerator implements SensorDataGenerator {
      * @param time The time of the measurement.
      * @return A measurement of carbon dioxide at the given position and time.
      */
-    public byte[] generateData(Integer x, Integer y, LocalTime time){
-        Double result = CarbonDioxideDataGenerator.generateData((double)x, (double)y);
+    public byte[] generateData(int x, int y, LocalTime time){
+        double result = CarbonDioxideDataGenerator.generateData(x, y);
         return new byte[]{(byte)Math.floorMod((int) Math.round(result),255)};
     }
     public byte[] generateData(Pair<Integer, Integer> pos, LocalTime time){
