+++ src/main/java/datagenerator/ParticulateMatterDataGenerator.java
@@ -17,7 +17,7 @@ public class ParticulateMatterDataGenerator implements SensorDataGenerator{
         random = new Random(DEFAULT_SEED);
     }
 
-    public static Double generateData(double x, double y) {
+    public static double generateData(double x, double y) {
         if (x < 250 && y < 250)
             return (double) 97 + (x + y) / 250 + 0.3 * random.nextGaussian();
         else if (x < 750 && y < 750)
@@ -38,8 +38,8 @@ public class ParticulateMatterDataGenerator implements SensorDataGenerator{
      * @param time The time of the measurement.
      * @return A measurement of particulate matter at the given position and time.
      */
-    public byte[] generateData(Integer x, Integer y, LocalTime time){
-        Double result = ParticulateMatterDataGenerator.generateData((double)x,(double)y);
+    public byte[] generateData(int x, int y, LocalTime time){
+        double result = ParticulateMatterDataGenerator.generateData(x, y);
         return new byte[]{(byte)Math.floorMod((int) Math.round(result),255)};
     }
     public byte[] generateData(Pair<Integer, Integer> pos, LocalTime time){
